+++ tcrypto/types/src/lib.rs
@@ -1,0 +1,7 @@
+// Copyright (c) 2022 The MobileCoin Foundation
+
+#![doc = include_str!("../README.md")]
+#![no_std]
+#![deny(missing_docs, missing_debug_implementations, unsafe_code)]
+
+pub mod ecdsa_p256;