+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `mc-sgx-core-types`: Idiomatic Rust Types for SGX primitives
+- `mc-sgx-tcrypto-types`: Idiomatic Rust Types for the `sgx_tcrypto` library
 - `mc-sgx-tservice-types`: Idiomatic Rust Types for the `sgx_tservice` library
 
 ### Changed